+++ apps/web/src/components/layout/app-shell/profile-client.tsx
@@ -1,48 +1,45 @@
 'use client'
 
-import Image from 'next/image'
 import { Popover, PopoverContent, PopoverPortal, PopoverTrigger } from '@radix-ui/react-popover'
 import Link from 'next/link'
 import s from './profile.module.scss'
 import { type User } from '@auth/core/types'
+import { User as UserIcon } from 'lucide-react'
 
-export function ProfileClient({ user }: { user?: Partial<User> }) {
+interface ProfileClientProps {
+  user?: Partial<User>
+  className?: string
+}
+
+function ProfileClient({ user, className }: ProfileClientProps) {
   if (!user) return null
-
-  const initials =
-    user?.name
-      ?.split(' ')
-      .map((word) => word[0])
-      .join('')
-      .toUpperCase() || '?'
-
   return (
-    <Popover>
-      <PopoverTrigger asChild>
-        <div className={s.profile}>
-          <div className={s.avatar}>
-            {user?.image ? (
-              <Image src={user.image} alt={user.name || 'User'} width={32} height={32} className={s.avatar} />
-            ) : (
-              <span className={s.label}>{initials}</span>
-            )}
+    <div className={`${className}`}>
+      <Popover>
+        <PopoverTrigger asChild>
+          <div className={s.profile}>
+            <div className={s.avatar}>
+              <UserIcon className='text-neutral-800 size-5' />
+            </div>
           </div>
-        </div>
-      </PopoverTrigger>
-      <PopoverPortal>
-        <PopoverContent collisionPadding={8} sideOffset={4}>
-          <div className={s.popover}>
-            <ul className={s.menu}>
-              <li>
-                <Link href='/profile'>Profile</Link>
-              </li>
-              <li>
-                <Link href='/api/auth/signout'>Sign out</Link>
-              </li>
-            </ul>
-          </div>
-        </PopoverContent>
-      </PopoverPortal>
-    </Popover>
+        </PopoverTrigger>
+        <PopoverPortal>
+          <PopoverContent collisionPadding={8} sideOffset={4}>
+            <div className={s.popover}>
+              <ul className={s.menu}>
+                <li>
+                  <Link href='/profile'>Profile</Link>
+                </li>
+                <li>
+                  <Link href='/api/auth/signout'>Sign out</Link>
+                </li>
+              </ul>
+            </div>
+          </PopoverContent>
+        </PopoverPortal>
+      </Popover>
+    </div>
   )
 }
+
+export default ProfileClient
