+++ apps/web/src/components/layout/app-shell/profile.module.scss
@@ -28,9 +28,9 @@
 }
 
 .popover {
-  background: var(--r-layer-01);
+  background: var(--r-layer);
   box-shadow: 0 2px 6px 0 rgba(0, 0, 0, 0.2);
-  padding: var(--spacing-2);
+  padding-left: 8px;
   border-radius: var(--radius-sm);
   width: 15rem;
 }
